% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brecha_gra.R
\name{brecha_gra}
\alias{brecha_gra}
\title{Grafica de brechas}
\usage{
brecha_gra(data_gap, xlab = NULL, ylab = NULL, leyenda = "Quantile")
}
\arguments{
\item{xlab}{String. Nombre del eje X de la gráfica. NULL por defecto.}

\item{ylab}{String. Nombre del eje Y de la gráfica. NULL por defecto.}

\item{leyenda}{String. Nombre de la leyenda. NULL por defecto.}

\item{data_ineq}{Data frame o Tibble. Datos obtenidos de la función 'data_proc' con la opción tipo = 'brecha'}
}
\description{
Cálculo de los coeficientes de una regresión lineal múltiple
}
\examples{
ineq_gra()
}
