% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brecha_tab.R
\name{brecha_tab}
\alias{brecha_tab}
\title{Grafica de brechas}
\usage{
brecha_tab(data_gap)
}
\arguments{
\item{data_ineq}{Data frame o Tibble. Datos obtenidos de la función 'data_proc' con la opción tipo = 'brecha'}
}
\description{
Cálculo de los coeficientes de una regresión lineal múltiple
}
\examples{
brecha_tab()
}
