% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{data_proc}
\alias{data_proc}
\title{Calculando índices para las gráficas y tablas}
\usage{
data_proc(
  data,
  tipo,
  periodo,
  vari.x,
  vari.ir,
  vari.num,
  anhos = "year",
  regiones = "country",
  poblacion = "population"
)
}
\arguments{
\item{data}{Data frame o Tibble. De contener regiones, población por regiones, años, una cantidad de casos e incidencia de alguna enfermedad, y alguna variable social}

\item{tipo}{Character. Define el tipo de procesamiento que se hará. 'desigualdad' o 'brecha'.}

\item{periodo}{Integer. Vector con los años que se utilizarán para el procesamiento.}

\item{vari.x}{Character. Nombre de la columna que tiene la determinante social.}

\item{vari.ir}{Character. Nombre de la columna con la incidiencia de alguna enfermedad.}

\item{vari.num}{Character. Nombre de la columna con la cantidad de casos de una enfermedad.}

\item{anhos}{Character. Nombre de la columna que tiene los años. 'year' por defecto.}

\item{regiones}{Character. Nombre de la columna con las regiones. 'country' por defecto.}

\item{poblacion}{Character. Nombre de la columan con la poblacion. 'population' por defecto.}
}
\description{
Cálculo de los coeficientes de una regresión lineal múltiple
}
\examples{
data_proc()
}
