% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reg.R
\name{data_reg}
\alias{data_reg}
\title{Funci贸n para generar las curvas de regresi贸n}
\usage{
data_reg(data_ineq)
}
\arguments{
\item{data_ineq}{Data frame o Tibble. Datos obtenidos de la funci贸n 'data_proc' con la opci贸n tipo = 'desigualdad'}
}
\description{
Permite estimar las curvas que se ajustan a la data social. Se calcula a partir de
la data de desigualdades.
}
\examples{
data_reg()
}
