% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desi_app.R
\name{desi_app}
\alias{desi_app}
\title{App para el analisis de las desigualdades y brechas sociales en salud.}
\usage{
desi_app(data, salud, social, years, years_sel = years[1])
}
\arguments{
\item{data}{Data frame o Tibble. De contener regiones, población por regiones, años, una cantidad de casos e incidencia de alguna enfermedad, y alguna variable social. La variable de años debe llamarse "year", la población "population" y la de regiones "country"}

\item{salud}{List. Lista con nombres genéricos respecto a diversos parámetros de salud. Cada elemento de la lista conténdrá dos parámetros: "IR" hace referencia a la variable en "data" que contenga la incidencia de una efermedad; mienrtas que "Num" será la variable con la cantidad total de casos.}

\item{social}{Vector. Vector con las variables de "data" que representen alguna característica social.}

\item{years}{Integer. Vector que contiene todo los años a analizar.}

\item{years_sel}{Integer. Vector con los años que serán seleccionados al iniciar el aplicativo. Por defecto, se seleccionará el primer elemnto de "years".}
}
\description{
Cálculo de los coeficientes de una regresión lineal múltiple
}
\examples{
data_proc()

# Variables sociales
social = c(
   'Esperanza de vida' = 'lexp',
   'Gasto en salud' = 'hexp',
   'Access to improved sanitation facilities' = 'impsfac',
   'TB detection rate' = 'pnotified')
# Variables de salud
salud = list(
   'Tuberculosis' = c(IR = 'ir_tb', Num = 'num_tb_cases')
   )

# Años
years = c(2000, 2005, 2010)

desi_app(ineq_data, salud, social, years)
}
