% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq_reg.R
\name{ineq_reg}
\alias{ineq_reg}
\title{Grafica de desigualdades - versión de regresion}
\usage{
ineq_reg(data_ineq, xlab = NULL, ylab = NULL, leyenda = "Años")
}
\arguments{
\item{data_ineq}{Data frame o Tibble. Datos obtenidos de la función 'data_proc' con la opción tipo = 'desigualdad'}

\item{xlab}{String. Nombre del eje X de la gráfica. NULL por defecto.}

\item{ylab}{String. Nombre del eje Y de la gráfica. NULL por defecto.}

\item{leyenda}{String. Nombre de la leyenda. NULL por defecto.}
}
\description{
Cálculo de los coeficientes de una regresión lineal múltiple
}
\examples{
ineq_reg()
}
