% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineq_tab.R
\name{ineq_tab}
\alias{ineq_tab}
\title{Grafica de brechas}
\usage{
ineq_tab(data_ineq)
}
\arguments{
\item{data_ineq}{Data frame o Tibble. Datos obtenidos de la función 'data_proc' con la opción tipo = 'desigualdad'.}
}
\description{
Cálculo de los coeficientes de una regresión lineal múltiple
}
\examples{
ineq_tab()
}
