% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_semisum.R
\name{lag_semisum}
\alias{lag_semisum}
\title{Promedio desfasado}
\usage{
lag_semisum(x)
}
\arguments{
\item{x}{Vector tipo numeric o integer}
}
\description{
Esta función realiza una semisuma con el dato en una posición anterior
}
\examples{
lag_semisum()
}
\keyword{semisum}
