% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_coef.R
\name{lm_coef}
\alias{lm_coef}
\title{Coeficientes de regresión lineal}
\usage{
lm_coef(y, x1, x2)
}
\arguments{
\item{y}{Vector que representa una variable de respuesta}

\item{x1}{Vector que representa la primera variable dependiente}

\item{x2}{Vector que representa la segunda variable dependiente}
}
\value{
Coeficientes de la regresión de \code{y} con \code{x1} y \code{x2}
}
\description{
Coeficientes de regresión lineal
}
\examples{
y <- rnorm(100)
x1 <- rnorm(100)
x2 <- rnorm(100)
lm_coef(y, x1, x2)
}
